Move packing logic to hf_dataset_packed_sequence.py
